fix: remaining bugs — ADC8x 1023→1024, ADC fallback mask, ADC8x consume clears intflags, Timer10 u64, Timer8 printf
N1 — ADC8x uses 1023 instead of 1024 in conversion (adc8x.cpp:295)
N2 — ADC classic fallback mask still 0x0F, H1 fix never committed (adc.cpp:301)
N3 — ADC8x consume_interrupt_request never clears intflags_ (adc8x.cpp:240-245)
N4 — Timer10 cycle_accumulator_ u32 wraps after ~268s (timer10.hpp:69)
N5 — Timer8 leftover debug printf floods stdout (timer8.cpp:314)

Also includes previous session fixes: AC8x absolute volts, M38 TWI slave,
M42 TWI8X bits_left 8→9, M46 USB uerst_ auto-clear, M47 USB data toggle PID,
CCL/EVSYS test fix, GPIO consume_pin_change fix.

All 182 tests pass.

+++ include/vioavr/core/timer10.hpp
@@ -66,7 +66,7 @@ private:
     MemoryBus* bus_ {};
     Adc* adc_ {};
 
-    u32 cycle_accumulator_ {0};
+    u64 cycle_accumulator_ {0};
     bool pll_enabled_ {false};
     u32 pck_frequency_ {64000000}; 
     u32 clock_ratio_ {1};
